Introduce a -v flag and refactor a little.
